Fix setup bug; Document 05-affected-people better
Fix the setup bug where the functions to download packages sat in a file
that required those packages to already exist. Moved them to 01-setup.R,
and changes the setup process. Source 01-setup, then get the packages,
then source 02-geography.

Document 05-affected-people.R better, particularly with what each step
does, and adding in headings for the sections of the code.